restructure primary pages, initial component restructure
