+++ src/pages/CheckoutPage.jsx
@@ -1,0 +1,3 @@
+export default function CheckoutPage() {
+  return <>Ready to checkout!</>;
+}
